pyproject.toml
[build-system]
requires = ["setuptools>=67.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "polymer-nlp-extractor"
version = "0.3.0"
description = "Advanced NLP pipeline for extracting polymer-related entities from scientific literature"
authors = [{ name = "Dhanush Mallu", email = "dhanushm@umich.edu" }]
readme = "README.md"
requires-python = ">=3.8"
license = { text = "MIT" }
dependencies = [
    "fastapi>=0.95.0",
    "uvicorn[standard]>=0.21.0",
    "torch>=2.0.0",
    "transformers>=4.30.0",
    "tokenizers>=0.13.0",
    "datasets>=2.12.0",
    "evaluate>=0.4.0",
    "accelerate>=0.20.0",
    "pytorch-crf>=0.7.2",
    "seqeval>=1.2.2",
    "notebook>=6.5.0",
    "scikit-learn>=1.1.0",
    "spacy>=3.5.0",
    "pandas>=1.5.0",
    "numpy>=1.24.0",
    "beautifulsoup4>=4.12.0",
    "lxml>=4.9.0",
    "unidecode>=1.3.0",
    "nltk>=3.8.0",
    "matplotlib>=3.6.0",
    "seaborn>=0.12.0",
    "weasyprint",
    "reportlab",
    "pandas-stubs>=2.3.0",
    "flask>=2.2.0",
    "python-multipart>=0.0.6",
    "cryptography>=39.0.0",
    "appwrite>=4.0.0",
    "tqdm>=4.65.0",
    "python-dotenv>=1.0.0",
    "fuzzywuzzy>=0.18.0",
    "python-Levenshtein>=0.21.0",
    "ipywidgets>=8.0.0",
    "ipyaggrid>=0.3.0",
    "wandb>=0.15.0",
    "grobid-client-python>=0.0.11",
    "requests",
    "psutil",
    "pyyaml>=6.0.0",
    "jsonschema>=4.17.0",
    "pytest>=7.0.0",
    "black>=23.0.0",
    "flake8>=6.0.0",
    "mypy>=1.2.0",
    "pre-commit>=3.0.0",
    "click>=8.0.0",
    "typer>=0.9.0",
    "huggingface_hub>=0.16.0",
    "wkhtmltopdf>=0.2.0",
]

[project.optional-dependencies]
dev = ["pytest>=7.0.0", "black>=23.0.0", "flake8>=6.0.0", "mypy>=1.2.0", "pre-commit>=3.0.0"]

[tool.setuptools.packages.find]
where = ["polymer_extractor"]

[project.scripts]
pnlp = "polymer_extractor.cli.cli:cli"
